Move calculateDbNameWithLimit to `utils/db`.
