+++ utils/db.js
@@ -1,0 +1,49 @@
+/**
+ * Copyright 2013-2020 the original author or authors from the JHipster project.
+ *
+ * This file is part of the JHipster project, see https://www.jhipster.tech/
+ * for more information.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+const crypto = require('crypto');
+
+/**
+ * get a constraint name for tables in JHipster preferred style after applying any length limits required.
+ *
+ * @param {string} entityName - name of the entity
+ * @param {string} columnOrRelationName - name of the column or related entity
+ * @param {object} options
+ * @param {boolean} options.noSnakeCase - do not convert names to snakecase
+ * @param {string} options.prefix
+ * @param {string} options.separator
+ * @param {boolean} options.appendHash - adds a calculated hash based on entityName and columnOrRelationName to prevent trimming conflict.
+ */
+function calculateDbNameWithLimit(entityName, columnOrRelationName, limit, options) {
+    const { noSnakeCase, prefix, separator, appendHash } = options;
+    const halfLimit = Math.floor(limit / 2);
+    const suffix = !appendHash
+        ? ''
+        : `_${crypto.createHash('shake256', { outputLength: 1 }).update(`${entityName}.${columnOrRelationName}`, 'utf8').digest('hex')}`;
+
+    let entityTable = noSnakeCase ? entityName : this.getTableName(entityName);
+    let otherTable = noSnakeCase ? columnOrRelationName : this.getTableName(columnOrRelationName);
+
+    entityTable = entityTable.substring(0, halfLimit - (!appendHash ? 0 : separator.length));
+    otherTable = otherTable.substring(0, limit - entityTable.length - separator.length - prefix.length - suffix.length);
+
+    return `${prefix}${entityTable}${separator}${otherTable}${suffix}`;
+}
+
+module.exports = { calculateDbNameWithLimit };
